+++ airflow-core/tests/unit/models/test_dagbag.py
@@ -36,7 +36,7 @@ from sqlalchemy import select
 
 from airflow import settings
 from airflow.models.dag import DAG, DagModel
-from airflow.models.dagbag import DagBag
+from airflow.models.dagbag import DagBag, _capture_with_reraise
 from airflow.models.dagwarning import DagWarning, DagWarningType
 from airflow.models.serialized_dag import SerializedDagModel
 from airflow.serialization.serialized_objects import SerializedDAG
@@ -1012,3 +1012,59 @@ with airflow.DAG(
             mock_signal.side_effect = ValueError("Invalid signal setting")
             DagBag(dag_folder=os.fspath(tmp_path), include_examples=False)
             assert "SIGSEGV signal handler registration failed. Not in the main thread" in caplog.text
+
+
+class TestCaptureWithReraise:
+    @staticmethod
+    def raise_warnings():
+        warnings.warn("Foo", UserWarning, stacklevel=2)
+        warnings.warn("Bar", UserWarning, stacklevel=2)
+        warnings.warn("Baz", UserWarning, stacklevel=2)
+
+    def test_capture_no_warnings(self):
+        with warnings.catch_warnings():
+            warnings.simplefilter("error")
+            with _capture_with_reraise() as cw:
+                pass
+            assert cw == []
+
+    def test_capture_warnings(self):
+        with pytest.warns(UserWarning, match="(Foo|Bar|Baz)") as ctx:
+            with _capture_with_reraise() as cw:
+                self.raise_warnings()
+            assert len(cw) == 3
+        assert len(ctx.list) == 3
+
+    def test_capture_warnings_with_parent_error_filter(self):
+        with warnings.catch_warnings(record=True) as records:
+            warnings.filterwarnings("error", message="Bar")
+            with _capture_with_reraise() as cw:
+                with pytest.raises(UserWarning, match="Bar"):
+                    self.raise_warnings()
+            assert len(cw) == 1
+        assert len(records) == 1
+
+    def test_capture_warnings_with_parent_ignore_filter(self):
+        with warnings.catch_warnings(record=True) as records:
+            warnings.filterwarnings("ignore", message="Baz")
+            with _capture_with_reraise() as cw:
+                self.raise_warnings()
+            assert len(cw) == 2
+        assert len(records) == 2
+
+    def test_capture_warnings_with_filters(self):
+        with warnings.catch_warnings(record=True) as records:
+            with _capture_with_reraise() as cw:
+                warnings.filterwarnings("ignore", message="Foo")
+                self.raise_warnings()
+            assert len(cw) == 2
+        assert len(records) == 2
+
+    def test_capture_warnings_with_error_filters(self):
+        with warnings.catch_warnings(record=True) as records:
+            with _capture_with_reraise() as cw:
+                warnings.filterwarnings("error", message="Bar")
+                with pytest.raises(UserWarning, match="Bar"):
+                    self.raise_warnings()
+            assert len(cw) == 1
+        assert len(records) == 1
